Fix and re-enable important compiler warnings
The disabled warnings have a high signal-to-noise ratio and would have
caught two bugs.

linux_set_iface_flags() had an error handling path that returned an
uninitalized ret to the caller.  This is fixed by setting ret
appropriately before returning it.

onSignificantWifiChange() made an implicit copy of an object, then tried
to read from a flexible array member in the copy.  This is fixed by
turning the implicit copy into a reference.

Change-Id: Ie1335b80a62c349ef766e3e31221ce87f72f6c0d
Signed-off-by: Greg Hackmann <ghackmann@google.com>